# contracts/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Erp.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/masters/parts": {
      "get": {
        "tags": [
          "Masters"
        ],
        "summary": "List parts",
        "description": "Server-paged. Supports sort=field:asc|desc (comma-separated), filter=field:op:value (semicolon-separated), and free-text search across part number, item code, description and HSN code. pageSize is clamped to 200.",
        "operationId": "ListParts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfPartListItemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfPartListItemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfPartListItemDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Masters"
        ],
        "summary": "Create a part",
        "description": "Creates the part in Draft status. It becomes usable once approved.",
        "operationId": "CreatePart",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePartRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePartRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/masters/parts/{id}": {
      "get": {
        "tags": [
          "Masters"
        ],
        "summary": "Get a part",
        "operationId": "GetPartById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PartDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Masters"
        ],
        "summary": "Update a part",
        "description": "Requires the rowVersion returned by GET. A stale value yields 409 rather than overwriting a concurrent edit. The part number cannot be changed here.",
        "operationId": "UpdatePart",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePartRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePartRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/masters/parts/{id}/approve": {
      "post": {
        "tags": [
          "Masters"
        ],
        "summary": "Approve a part",
        "description": "Approves a part awaiting review. The author cannot approve their own part.",
        "operationId": "ApprovePart",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/masters/parts/{id}/submit": {
      "post": {
        "tags": [
          "Masters"
        ],
        "summary": "Submit a part for approval",
        "description": "Moves a Draft part to PendingApproval. It cannot be edited while under review.",
        "operationId": "SubmitPartForApproval",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/masters/parts/import": {
      "post": {
        "tags": [
          "Masters"
        ],
        "summary": "Import parts from an Excel file",
        "description": "Accepts a single .xlsx upload of at most 5,000 rows. All or nothing: every row is parsed and checked first, and if anything is wrong nothing is written and the response is 422 with every problem in the file. On success the response is 200 and the report says how many rows landed.",
        "operationId": "Importparts",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResultDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResultDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResultDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResultDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResultDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/masters/parts/import-template": {
      "get": {
        "tags": [
          "Masters"
        ],
        "summary": "Download the parts import template",
        "description": "An empty workbook whose headings are exactly what the import expects, plus a 'Column guide' sheet describing each column. Generated from the same column definitions the importer parses, so the two cannot drift.",
        "operationId": "GetPartsImportTemplate",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePartRequest": {
        "required": [
          "partNumber",
          "description",
          "unitOfMeasureCode"
        ],
        "type": "object",
        "properties": {
          "partNumber": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "categoryId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "unitOfMeasureCode": {
            "type": "string"
          },
          "hsnCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "drawingNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "attributes": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PartAttributesDto"
              }
            ]
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
      "ImportResultDto": {
        "required": [
          "master",
          "totalRows",
          "importedRows",
          "committed",
          "errors",
          "errorsTruncated"
        ],
        "type": "object",
        "properties": {
          "master": {
            "type": "string"
          },
          "totalRows": {
            "type": "integer",
            "format": "int32"
          },
          "importedRows": {
            "type": "integer",
            "format": "int32"
          },
          "committed": {
            "type": "boolean"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportRowErrorDto"
            }
          },
          "errorsTruncated": {
            "type": "boolean"
          }
        }
      },
...
      "PagedResultOfPartListItemDto": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PartListItemDto"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        }
      },
...
      "PartDetailDto": {
        "required": [
          "id",
          "partNumber",
          "description",
          "unitOfMeasureCode",
          "attributes",
          "isActive",
          "status",
          "businessUnitId",
          "createdAtUtc",
          "rowVersion"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "partNumber": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "categoryId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "unitOfMeasureCode": {
            "type": "string"
          },
          "hsnCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "drawingNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "attributes": {
            "$ref": "#/components/schemas/PartAttributesDto"
          },
          "isActive": {
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/PartStatusDto"
          },
          "businessUnitId": {
            "type": "integer",
            "format": "int32"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "rowVersion": {
            "type": "string"
          }
        }
      },
      "PartListItemDto": {
        "required": [
          "id",
          "partNumber",
          "originalPartNumber",
          "description",
          "unitOfMeasureCode",
          "isActive",
          "status",
          "createdAtUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "partNumber": {
            "type": "string"
          },
          "originalPartNumber": {
            "type": "string"
          },
          "itemNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": "string"
          },
          "technicalSpecification": {
            "type": [
              "null",
              "string"
            ]
          },
          "moc": {
            "type": [
              "null",
              "string"
            ]
          },
          "partCategoryCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "partType": {
            "type": [
              "null",
              "string"
            ]
          },
          "formCategory": {
            "type": [
              "null",
              "string"
            ]
          },
          "unitOfMeasureCode": {
            "type": "string"
          },
          "purchaseUomCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "sellingUomCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "materialType": {
            "type": [
              "null",
              "string"
            ]
          },
          "seriesCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "partRevisionNo": {
            "type": [
              "null",
              "string"
            ]
          },
          "sourceCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "weightKg": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "leadTimeDays": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "minimumStockLevel": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "reorderPoint": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "hsnCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "drawingNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/PartStatusDto"
          },
          "revisionRemark": {
            "type": [
              "null",
              "string"
            ]
          },
          "holdRemark": {
            "type": [
              "null",
              "string"
            ]
          },
          "inactiveRemark": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdBy": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedBy": {
            "type": [
              "null",
              "string"
            ]
          },
          "modifiedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "PartStatusDto": {
        "enum": [
          "Draft",
          "PendingApproval",
          "Approved",
          "Rejected",
          "Hold"
        ]
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UpdatePartRequest": {
        "required": [
          "description",
          "unitOfMeasureCode",
          "rowVersion"
        ],
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "categoryId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "unitOfMeasureCode": {
            "type": "string"
          },
          "hsnCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "drawingNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "attributes": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PartAttributesDto"
              }
            ]
          },
          "rowVersion": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}